Split gql client in service/indexer parts.
