+++ linera-indexer-graphql-client/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "linera-indexer-graphql-client"
+version = "0.4.0"
+edition = "2021"
+description = "GraphQL client for the indexer"
+readme = "README.md"
+repository = "https://github.com/linera-io/linera-protocol"
+homepage = "https://linera.dev"
+documentation = "https://docs.rs/linera-indexer-graphql-client/latest/linera_indexer_graphql_client/"
+license = "Apache-2.0"
+
+[dependencies]
+graphql_client = { version = "0.13", features = [ "reqwest" ] }
+linera-base = { workspace = true }
+serde = { workspace = true, features = [ "derive" ] }
+serde_json = { workspace = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+linera-execution = { workspace = true }
+
+[dev-dependencies]
+linera-service = { workspace = true }
+test-log = { workspace = true, features = ["trace"] }
+tempfile = {workspace = true }
+tokio = { workspace = true, features = ["full", "test-util"] }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true, features = ["fmt"] }
