+++ linera-indexer-graphql-client/gql/operations_requests.graphql
