+++ linera-indexer-graphql-client/src/indexer.rs
@@ -6,16 +6,16 @@ use linera_base::{crypto::CryptoHash, data_types::BlockHeight, identifiers::Chai
 
 #[derive(GraphQLQuery)]
 #[graphql(
-    schema_path = "graphql/indexer_schema.graphql",
-    query_path = "graphql/indexer_requests.graphql",
+    schema_path = "gql/indexer_schema.graphql",
+    query_path = "gql/indexer_requests.graphql",
     response_derives = "Debug, Serialize, Clone"
 )]
 pub struct Plugins;
 
 #[derive(GraphQLQuery)]
 #[graphql(
-    schema_path = "graphql/indexer_schema.graphql",
-    query_path = "graphql/indexer_requests.graphql",
+    schema_path = "gql/indexer_schema.graphql",
+    query_path = "gql/indexer_requests.graphql",
     response_derives = "Debug, Serialize, Clone"
 )]
 pub struct State;
