+++ linera-indexer-graphql-client/src/lib.rs
@@ -1,0 +1,7 @@
+// Copyright (c) Zefchain Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+//! A GraphQL client for the indexer.
+
+pub mod indexer;
+pub mod operations;
