+++ linera-indexer/Cargo.toml
@@ -29,7 +29,7 @@ linera-base = { workspace = true }
 linera-chain = { workspace = true }
 linera-core = { workspace = true }
 linera-execution = { workspace = true }
-linera-graphql-client = { workspace = true }
+linera-service-graphql-client = { workspace = true }
 linera-views = { workspace = true }
 reqwest = { workspace = true }
 serde = { workspace = true }
@@ -42,7 +42,9 @@ tracing-subscriber = { workspace = true, features = ["fmt"] }
 
 [dev-dependencies]
 anyhow = { workspace = true }
+linera-indexer-graphql-client = { workspace = true }
 linera-service = { workspace = true, features = ["rocksdb"] }
+linera-service-graphql-client = { workspace = true }
 once_cell = { workspace = true }
 test-log = { workspace = true, features = ["trace"] }
 serde_json = { workspace = true }
