+++ linera-indexer/tests/test.rs
@@ -2,13 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use linera_base::{data_types::Amount, identifiers::ChainId};
-use linera_graphql_client::{
+use linera_indexer_graphql_client::{
     indexer::{plugins, state, Plugins, State},
     operations::{get_operation, GetOperation, OperationKey},
-    request,
-    service::{block, transfer, Block, Transfer},
 };
 use linera_service::client::{resolve_binary, LocalNetwork, Network};
+use linera_service_graphql_client::{block, request, transfer, Block, Transfer};
 use once_cell::sync::Lazy;
 use std::{rc::Rc, str::FromStr, time::Duration};
 use tempfile::TempDir;
