+++ linera-service-graphql-client/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "linera-graphql-client"
+name = "linera-service-graphql-client"
 version = "0.4.0"
 edition = "2021"
-description = "GraphQL client for the node service and indexer"
+description = "A GraphQL client for Linera node service"
 readme = "README.md"
 repository = "https://github.com/linera-io/linera-protocol"
 homepage = "https://linera.dev"
-documentation = "https://docs.rs/linera-graphql-client/latest/linera_graphql_client/"
+documentation = "https://docs.rs/linera-service-graphql-client/latest/linera_service_graphql_client/"
 license = "Apache-2.0"
 
 [dependencies]
@@ -18,7 +18,7 @@ serde_json = { workspace = true }
 thiserror = { workspace = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-linera-chain = { workspace = true, features = ["test"] }
+linera-chain = { workspace = true }
 linera-core = { workspace = true }
 linera-execution = { workspace = true }
