+++ linera-service-graphql-client/README.md
@@ -1,0 +1,22 @@
+# Linera service GraphQL client
+
+<!-- cargo-rdme start -->
+
+A GraphQL client for the node service.
+
+<!-- cargo-rdme end -->
+
+## Generate schema
+
+To generate the linera service GraphQL schema, a binary `linera-export-schema` is available:
+```bash
+cargo run --bin linera-schema-export > linera-service-graphql-client/gql/service_schema.graphql
+```
+
+## Contributing
+
+See the [CONTRIBUTING](../CONTRIBUTING.md) file for how to help out.
+
+## License
+
+This project is available under the terms of the [Apache 2.0 license](../LICENSE).
