+++ linera-service-graphql-client/src/lib.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Zefchain Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+//! A GraphQL client for the node service.
+
+mod service;
+pub mod utils;
+
+pub use service::*;
+pub use utils::*;
